Add AG & decrypt logic
