+++ packages/web-app/app/_ui/components/index.tsx
@@ -3,3 +3,4 @@ export { Info } from './info';
 export { Topbar } from './topbar';
 export { Footer } from './footer';
 export { WalletButton } from './wallet-button';
+export { ProjectsTable } from './projects-table';
