+++ packages/web-app/app/_ui/components/info.tsx
@@ -1,134 +1,15 @@
 'use client';
 
-import { useAccount } from 'wagmi';
 import { Button } from './button';
 import { useIdOS } from '@/app/_providers/idos';
-import { useState } from 'react';
-import { idOS } from '@idos-network/idos-sdk';
-import { getAccessGrantsContentDecrypted } from '@/app/server/idos';
-import {
-  useFetchCredentials,
-  useFetchGrants,
-  useFetchWallets,
-} from '@/app/_lib/queries';
+import { IdosInfo } from './idos-info';
+import { PublicInfo } from '@/app/server/types';
 
-const json = (object) => {
-  return JSON.stringify(object, '', 2).replace(/"(data:.*?;).*/g, '$1 (...)');
+type TInfoProps = {
+  serverInfo: PublicInfo;
 };
 
-const DecryptButton = ({ dataId }: { dataId: string }) => {
-  const getAccessGrants = getAccessGrantsContentDecrypted.bind(null, dataId);
-
-  return (
-    <form action={getAccessGrants}>
-      <Button type="submit">Decrypt</Button>
-    </form>
-  );
-};
-
-const IdosInfo = () => {
-  const [contentStr, setContentStr] = useState('');
-  const { address } = useAccount();
-  const { sdk } = useIdOS();
-  const { data, isError, isLoading } = useFetchCredentials();
-  const { data: wallets } = useFetchWallets();
-  const { data: grants } = useFetchGrants();
-  // const { data: credentials } = useFetchCredentials();
-
-  return (
-    <div className="mt-2">
-      <h3>Idos</h3>
-      <h4 className="my-4 font-semibold">Wallets</h4>
-      {wallets?.map((wallet) => (
-        <div key={wallet.id}>
-          <p>
-            <span>Id: </span>
-            {wallet.id}
-          </p>
-          <p>
-            <span>Address: </span>
-            {wallet.address}
-          </p>
-        </div>
-      ))}
-      <h4 className="mt-6 font-semibold">Credentials</h4>
-      {data?.map((credential) => (
-        <div key={credential.id} className="flex flex-col gap-2 py-4">
-          <p>
-            <span>Id: </span>
-            <Button
-              className="p-2"
-              onClick={async () => {
-                if (!sdk) return;
-
-                const { content } = await sdk.data.get(
-                  'credentials',
-                  credential.id,
-                );
-
-                try {
-                  console.log(
-                    '%c==>Verify Credential:',
-                    'color: green; background: yellow; font-size: 20px',
-                    await idOS.verifiableCredentials.verify(content),
-                  );
-                } catch (error) {
-                  console.log(
-                    '%c==>',
-                    'color: green; background: yellow; font-size: 20px',
-                    error,
-                  );
-                }
-
-                try {
-                  const result = await sdk.grants.list({
-                    owner: address,
-                    // grantee: fakeServer.publicInfo.granteeAddress,
-                  });
-
-                  console.log(
-                    '%c==>GRANTS',
-                    'color: green; background: yellow; font-size: 20px',
-                    result,
-                  );
-                } catch (error) {}
-
-                setContentStr(JSON.parse(content));
-              }}
-            >
-              {credential.id}
-            </Button>
-          </p>
-          <p>
-            <span>Level: </span>
-            {credential.credential_level}
-          </p>
-          <p>
-            <span>Status: </span>
-            {credential.credential_status}
-          </p>
-        </div>
-      ))}
-      {contentStr ? <pre>${json(contentStr)}</pre> : null}
-      <h4 className="mt-6 font-semibold">Grants</h4>
-      {grants?.map((grant) => (
-        <div key={grant.dataId} className="flex flex-col gap-2 py-4">
-          <p>
-            <span>DataId: </span>
-            {grant.dataId}
-          </p>
-          <p>
-            <span>Owner: </span>
-            {grant.owner}
-          </p>
-          <DecryptButton dataId={grant.dataId} />
-        </div>
-      ))}
-    </div>
-  );
-};
-
-export function Info() {
+export function Info({ serverInfo }: TInfoProps) {
   const { hasSigned, authenticate, address } = useIdOS();
 
   if (!address) return null;
@@ -136,7 +17,7 @@ export function Info() {
   return (
     <div className="grid grid-cols-1 gap-2">
       {hasSigned ? (
-        <IdosInfo />
+        <IdosInfo serverInfo={serverInfo} />
       ) : (
         <Button onClick={authenticate}>Check KYC status</Button>
       )}
