+++ packages/web-app/app/server/types.ts
@@ -10,3 +10,9 @@ export interface Grant {
   content: string;
   encryption_public_key: string;
 }
+
+export interface PublicInfo {
+  grantee: string;
+  encryptionPublicKey: string;
+  lockTimeSpanSeconds: number;
+}
